refactor(geometry): Remove AMR and mesh duplicate files from root
Completes partial subdirectory migration for geometry module by removing
duplicate AMR and mesh files from root directory. Subdirectories are now
the canonical locations.

Changes:
- Update __init__.py to import from amr/ subdirectory (amr.amr_1d, etc.)
- Update 3 files importing AMR classes to use subdirectory paths
- Update base_environment.py to import Mesh2D from meshes/ subdirectory
- Remove 4 AMR duplicate files from root: amr_1d.py, amr_quadtree_2d.py,
  amr_triangular_2d.py, amr_tetrahedral_3d.py
- Remove 4 mesh duplicate files from root: mesh_1d.py, mesh_2d.py,
  mesh_3d.py, mesh_manager.py

Files updated:
- mfg_pde/geometry/__init__.py (AMR imports)
- examples/basic/geometry/amr_geometry_protocol_demo.py
- examples/advanced/geometry_advanced/amr_1d_geometry_demo.py
- docs/development/v0.10.1_amr_geometry_protocol.md
- mfg_pde/alg/reinforcement/multi_population/base_environment.py

Issue #329: Geometry module subdirectory cleanup
