+++ docs/development/v0.10.1_amr_geometry_protocol.md
@@ -159,7 +159,7 @@ assert config["legacy_1d_attrs"] is None  # AMR doesn't support legacy
 
 ```python
 from mfg_pde import ExampleMFGProblem
-from mfg_pde.geometry.amr_1d import OneDimensionalAMRMesh, AMRRefinementCriteria
+from mfg_pde.geometry.amr.amr_1d import OneDimensionalAMRMesh, AMRRefinementCriteria
 from mfg_pde.geometry.domain_1d import Domain1D
 from mfg_pde.geometry.boundary_conditions_1d import BoundaryConditions
