+++ examples/advanced/geometry_advanced/amr_1d_geometry_demo.py
@@ -28,7 +28,7 @@ import numpy as np
 
 from mfg_pde import ExampleMFGProblem
 from mfg_pde.factory import create_fast_solver
-from mfg_pde.geometry.amr_1d import AMRRefinementCriteria, OneDimensionalAMRMesh
+from mfg_pde.geometry.amr.amr_1d import AMRRefinementCriteria, OneDimensionalAMRMesh
 from mfg_pde.geometry.domain_1d import Domain1D
 from mfg_pde.utils.mfg_logging import configure_research_logging, get_logger
