+++ examples/basic/geometry/amr_geometry_protocol_demo.py
@@ -20,8 +20,8 @@ Part of: v0.10.1 AMR GeometryProtocol support
 import matplotlib.pyplot as plt
 import numpy as np
 
-from mfg_pde.geometry.amr_1d import AMRRefinementCriteria, OneDimensionalAMRMesh
-from mfg_pde.geometry.amr_quadtree_2d import AdaptiveMesh
+from mfg_pde.geometry.amr.amr_1d import AMRRefinementCriteria, OneDimensionalAMRMesh
+from mfg_pde.geometry.amr.amr_quadtree_2d import AdaptiveMesh
 from mfg_pde.geometry.boundary_conditions_1d import BoundaryConditions
 from mfg_pde.geometry.domain_1d import Domain1D
 from mfg_pde.geometry.geometry_protocol import (
